# Optimize box-shadow rendering (use segments for pictures).