For my HD4600 GPU, this drops the time spent in the
many-boxshadows.html benchmark from 8.3 ms to 1.6ms. It
shows similar wins on the pocket page in Gecko in #1993.

Enable segment rendering for box-shadows (and Pictures in
general). This removes a large number of alpha-blended
pixels from outset box shadows, by cutting out segments
that have ClipOut mode.

Also modify the edge AA flags support, and how segment builder
produces these. The segment builder marks any edge that is on
the primitive edge, or cut out by a ClipOut rect as needing
AA, meaning we get smooth AA on internal (clipped out) edges.

Fixes #1993 (this page runs at ~5ms on my HD4600 now, easily
holding 60 fps).

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/servo/webrender/2311)
<!-- Reviewable:end -->