@glennw 2018-01-16
r? @kvark

The way that the brush / picture code interacts is a bit untidy. However, since this is such a significant performance win, I think it's probably reasonable to review / merge this, and then to the refactoring of the code as a follow up once I think about it a bit more.

I still need to do a gecko try run for this - I wouldn't be surprised if there are a few failures due to changing some of the fiddly AA-related code :)