@glennw 2018-01-17
Gecko try seems good, I think (but could do with a sanity check to make sure I didn't miss something):

https://hg.mozilla.org/try/rev/0faba5f8ff906003d40dd7f1b4308ae8ed8897b3
https://treeherder.mozilla.org/#/jobs?repo=try&revision=0faba5f8ff906003d40dd7f1b4308ae8ed8897b3

```
9 - unrelated JS intermittent
gl3 - unexpected PASS (intermittent?)
R2 - 4 new PASS (from #2292)
R4 - 1 new FAIL (from #2285)
R8 - 3 new FAIL (from #2292)
```