@glennw 2018-01-17
This also seems to have a good effect of a few of the other existing bugs with slow reported GPU times. But I'll run through those once this lands in Gecko and close as appropriate.