@staktrace 2018-01-17
> Gecko try seems good, I think (but could do with a sanity check to make sure I didn't miss something):

Your assessment seems correct - the only one I'm not sure of is the gl3 unexpected PASS. It doesn't look like a pre-existing issue (there is a related bug it apprarently it was fixed already). I've retriggered that job a few more times on your try push to see if it's consistent or intermittent.