@glennw 2018-01-17
@staktrace Yea, it does seem unlikely to be an intermittent since it passed on both the debug and release builds. Looking at the test itself, it doesn't look like it actually draws anything - I think it just creates a webgl context and queries an extension. In which case, I don't know what could have changed to make that pass... :man_shrugging: