@kvark 2018-01-17
Good stuff!
I can see traces of https://github.com/servo/webrender/pull/2263/commits/8804cab68354572d7711cf2c48999c158d358287 here, and the new edge segmentation logic is precisely what I was missing (so hopefully will come back to that PR soon).

---

Reviewed 10 of 10 files at r1.
Review status: all files reviewed at latest revision, 5 unresolved discussions.

---

*[webrender/src/batch.rs, line 1323 at r1](https://reviewable.io:443/reviews/servo/webrender/2311#-L33ah3oGRq24lFuP_Tm:-L33ah3ps6zNhZOHUuiX:b-qxvvvo) ([raw file](https://github.com/servo/webrender/blob/9b939a4061b23181e4672116ecd8f4f098eed413/webrender/src/batch.rs#L1323)):*
> ```Rust
>             }
>             BrushKind::Picture => {
>                 panic!("bug: get_batch_key is handled at higher level for pictures");
> ```

I see what you mean by the ugliness of brush-picture interaction now :)

---

*[webrender/src/box_shadow.rs, line 258 at r1](https://reviewable.io:443/reviews/servo/webrender/2311#-L33anROM4AotmJLFIMK:-L33anRQkKi1Cj_39c7b:bjwa7hh) ([raw file](https://github.com/servo/webrender/blob/ef420106c1d28a7f1b342201b837d4b5f1a53831/webrender/src/box_shadow.rs#L258)):*
> ```Rust
>                     );
> 
>                     // TODO(gw): Right now, we always use a clip out
> ```

:tada: 

---

*[webrender/src/prim_store.rs, line 305 at r1](https://reviewable.io:443/reviews/servo/webrender/2311#-L33b8P-zxxlx7IbVch1:-L33b8P-zxxlx7IbVch2:b-sbt41) ([raw file](https://github.com/servo/webrender/blob/9b939a4061b23181e4672116ecd8f4f098eed413/webrender/src/prim_store.rs#L305)):*
> ```Rust
>         // has to match VECS_PER_SPECIFIC_BRUSH
>         match self.kind {
>             BrushKind::Picture => {
> ```

this doesn't match the comment on top any more
perhaps, we can call the `self.brush.write_gpu_blocks` here?

---

*[webrender/src/segment.rs, line 435 at r1](https://reviewable.io:443/reviews/servo/webrender/2311#-L33cNQd96F4BqdT6VPC:-L33bv1soumCrFqorFHz:bq9xlic) ([raw file](https://github.com/servo/webrender/blob/9b939a4061b23181e4672116ecd8f4f098eed413/webrender/src/segment.rs#L435)):*
> ```Rust
>                 let mut edge_flags = EdgeAaSegmentMask::empty();
> 
>                 if x == 0 || segments[y * x_count + x - 1].is_none() {
> ```

interesting! could probably be refactored quite a bit (bit masks, not bothering computing the `edge_flags` if the current is `None`, etc), but later

---

*[webrender/src/segment.rs, line 585 at r1](https://reviewable.io:443/reviews/servo/webrender/2311#-L33ceijIMT14EnuWsP4:-L33ceijIMT14EnuWsP5:bck8hmz) ([raw file](https://github.com/servo/webrender/blob/9b939a4061b23181e4672116ecd8f4f098eed413/webrender/src/segment.rs#L585)):*
> ```Rust
>         sb.build(|segment| {
>             segments.push(Segment {
>                 ..*segment
> ```

mmm that one seems odd. Couldn't we just do `segments.push(*segment)`?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/servo/webrender/2311#-:-L33d5fAUea3BM4x-7hc:bxoc9xx)*
<!-- Sent from Reviewable.io -->