@bors-servo 2018-01-17
:pushpin: Commit 9b939a4 has been approved by `kvark`

<!-- @bors-servo r=kvark 9b939a4061b23181e4672116ecd8f4f098eed413 -->