@staktrace 2018-01-17
The retriggers are consistently showing the same result for gl3. I would try doing another try push based on the latest m-c. Sometimes things like this go away with a new base revision. If it still happens we'll probably need to investigate more.