@bors-servo 2018-01-17
:sunny: Test successful - [status-appveyor](https://ci.appveyor.com/project/servo/webrender/build/1.0.7520), [status-taskcluster](https://tools.taskcluster.net/task-group-inspector/#/ejZQsFi0Qm-AZqIEOsQQaw)
Approved by: kvark
Pushing e5514a40a8823a9a90fcf9a0e18067772c53d29e to master...