@glennw 2018-01-17
@staktrace OK, I'll do a try run with this today and investigate further if it still passes.